Add files via upload
This is an update to the first commit done by Mathew Nann. This code opens the website and its separate routes with the link that is printed into the terminal. There are some HTML code in the Template folder. This HTML code is the design/structure of the website, and this structure has been copied from Tech with Tim's Youtube video (https://www.youtube.com/watch?v=dam0GPOAvVI&ab_channel=TechWithTim).